updates based on comments
