+++ package.json
@@ -45,7 +45,5 @@
     ]
   },
   "devDependencies": {
-    "dotenv-webpack": "^8.0.1",
-    "react-error-overlay": "^6.0.9"
   }
 }
