+++ src/components/common/Map.tsx
@@ -3,9 +3,7 @@ import 'components/common/Map.css'
 
 const Map = () => {
   return(
-    <div>
       <img src={map} alt="map" className="home--map" />
-    </div>
   )
 }
