+++ src/index.tsx
@@ -1,11 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import SignUp from 'components/SignUp';
-import SignIn from 'components/SignIn';
-import Welcome from 'components/Welcome';
+
+import App from 'components/App'
 import reportWebVitals from './reportWebVitals';
-import Paths from 'constants/paths';
-import { BrowserRouter, Routes, Route } from "react-router-dom";
+import { BrowserRouter } from "react-router-dom";
 import './index.css';
 import 'components/common/general.css'
 
@@ -15,11 +13,7 @@ const root = ReactDOM.createRoot(
 root.render(
   <React.StrictMode>
     <BrowserRouter>
-      <Routes>
-        <Route path={Paths.ROOT} element={<SignIn />} />
-        <Route path={Paths.SIGN_UP} element={<SignUp />} />
-        <Route path={Paths.WELCOME} element={<Welcome />} />
-      </Routes>
+      <App />
     </BrowserRouter>
   </React.StrictMode>
 );
